Implementa gestión de inventario y movimientos en la aplicación
- Se añadieron nuevos permisos relacionados con la gestión de inventario y movimientos, incluyendo "LEER_INVENTARIO", "CREAR_INVENTARIO", "EDITAR_INVENTARIO", "ELIMINAR_INVENTARIO", "ASIGNAR_STOCK", "LEER_MOVIMIENTOS", "CREAR_MOVIMIENTOS" y "VER_REPORTES_INVENTARIO".
- Se actualizó la configuración de seguridad para permitir el acceso a las nuevas rutas de inventario y movimientos según los permisos asignados.
- Se implementaron nuevos métodos en los controladores para gestionar la asignación de stock y la visualización de movimientos de inventario, mejorando la funcionalidad de la aplicación.
- Se realizaron ajustes en las plantillas HTML para incluir opciones de gestión de stock y visualización de movimientos, mejorando la experiencia del usuario.

+++ src/main/java/com/diedari/jimdur/service/UbicacionService.java
@@ -6,8 +6,8 @@ import com.diedari.jimdur.model.Ubicaciones;
 
 public interface UbicacionService {
     public List<Ubicaciones> listarUbicaciones();
-    public Ubicaciones obtenerUbicacionPorId(Integer idUbicacion);
+    public Ubicaciones obtenerUbicacionPorId(Long idUbicacion);
     public Ubicaciones guardarUbicacion(Ubicaciones ubicacion);
-    public void eliminarUbicacion(Integer idUbicacion);
+    public void eliminarUbicacion(Long idUbicacion);
     public Ubicaciones actualizarUbicacion(Ubicaciones ubicacion);
 }
